% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nim_occupancy.R
\name{nimble.occ}
\alias{nimble.occ}
\title{Estimation of Single-Season Occupancy Models}
\usage{
nimble.occ(siteformula = NULL, obsformula = NULL, y = NULL,
  sitevars = NULL, obsvars = NULL, priors = c("Normal", "t", "Uniform",
  "Gamma"), dropbase = TRUE, droplast = FALSE, niter = 10000,
  burnin = 1000, initmcmc = c(0.5, 1, 5, 10), chains = 4,
  returncode = TRUE, returnsamp = TRUE, ...)
}
\arguments{
\item{siteformula}{A linear model formula for the latent occupancy state: \eqn{logit(\psi_{i})}.  Formula must start with a tilde (~).  Add 1 for an intercept.
Random effects are specified with a pipe (e.g., (1|A)).  Interactions can be specified with * (e.g., x*y).}

\item{obsformula}{A linear model formula for the observation process: \eqn{logit(p_{i,j})}.}

\item{y}{Detection and non detection data.  Must be provided as either a matrix or dataframe. Rows are sites and columns are replicates.  1 is detection and 0 is non detection.}

\item{sitevars}{A dataframe of site-specific covariates.  Must be a dataframe with named columns corresponding to variables.  Number of rows should equal the number of sites.}

\item{obsvars}{A named list with site and survey specific covariates.  Each list element is a matrix with rows corresponding to site and columns to replicate.
Variables specific in model must match the names in the list.}

\item{priors}{Prior options for all model parameters.  One prior is assigned to all model parameters specified in the siteformula and obsformula.  The default is a Normal(mu = 0, sd = 1000).  Other options are
t(mu = 0, tau = 1000, df = 5), Uniform(0, 1000), and Gamma(shape = .001, rate = .001).}

\item{dropbase}{Drops the first level of a categorical variable.  Default is TRUE.}

\item{droplast}{Drops the last level of a categorical variable.  Default is FALSE.}

\item{niter}{Number of iterations to run the MCMC.  Default is 10,000 iterations.}

\item{burnin}{Number of burnin iterations for MCMC.  Default is 1000 iterations.}

\item{initmcmc}{Vector of initial values for MCMC.  Length of vector should equal the number of chains.  Default is c(1, 5, 10, 15)}

\item{chains}{Number of MCMC chains to run.  Default is 4.}

\item{returncode}{Option to return the BUGS code passed to NIMBLE.  Default is TRUE.}

\item{returnsamp}{Option to return all posterior draws.  Default is TRUE}

\item{...}{User can specify different parameters for the priors.  Options are location, scale, lb, ub, df, shape, rate.}
}
\value{
The output will be a named list with the following elements: Summary, BUGScode, and Samples.  In the Summary statement, variables from the siteformula will start with "s."
and variables in the obsformula will start with "o."  Also, note parameters are on the transformed scale (log for siteformula variables; logit for obsformula variables).
In addition to quantiles, the effective sample size and Gelman Rubin diagnoistic are provided from the coda package.
}
\description{
\code{nimble.occ} returns generated BUGS code, model summary statistics, and posterior draws
}
\details{
This is a function for fitting Single Season-Occupany Models using a Bayeisan framework.  A basic structure of the type of model is:
\deqn{z_{i} ~ Bernoulli(\psi_{i})}
\deqn{y_{i,j} | z_{i} ~ Bernoulli(z_{i} * p_{i,j})}
Where i is the number of sites and j the number of temporal replicates. z is a binary latent variable for the true observed state(occupied or not occupied).
  Both \eqn{\psi} and p can be modeled as a linear function of covariates via a linked GLM.  The link for \eqn{\psi} and p is the logit link.
 Estimation is performed using NIMBLE.  In addition to fitting the model, generated BUGS code for the full model,
  as well as posterior draws are available.
}
\examples{
# Simualte Data from Kery and Schaub (p. 421): \\url{http://www.sciencedirect.com/science/book/9780123870209}
R <- 200
T <- 3
psi <- 0.8
p <- 0.5
z <- rbinom(n = R, size = 1, prob = psi)
y <- matrix(NA, nrow = R, ncol = T)
for (i in 1:T){y[,i] <- rbinom(n = R, size = 1, prob = z*p)}



#Fit Model
model <- nimble.occ(siteformula = ~ 1, obsformula = ~ 1, y = y, initmcmc = 1, chains = 1)


}
\author{
Colin Lewis-Beck
}
