% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nimbleEcology.R
\docType{package}
\name{nimbleEcology}
\alias{nimbleEcology}
\alias{nimbleEcology-package}
\title{nimbleEcology: A package for fitting Ecological Models using NIMBLE}
\description{
The nimbleEcology package currently fits the following classes of models:
Abundance Models (Binomial Mixture Models), Single Season Occupancy Models, and Multiseason Occupancy Models.
}
\section{main nimbleEcology functions}{

 \itemize{
\item{nimble.abun}
\item{nimble.occ}
\item{nimble.dynamic.occ}
}
}

